Code/Optuna/optuna_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex==0.1==conda_forge
_openmp_mutex=4.5=2_gnu
_tflow_select=2.3.0=mkl
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.3=py310h5764c6d_1
aiosignal=1.3.1=pyhd8ed1ab_0
alembic=1.8.1=pyhd8ed1ab_0
alsa-lib=1.2.3.2=h166bdaf_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=py310h06a4308_0
attrs=22.1.0=pyh71513ae_1
autopage=0.5.1=pyhd8ed1ab_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=openblas
blinker=1.5=pyhd8ed1ab_0
blosc=1.21.0=h4ff587b_1
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.9.24=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h74dc2b5_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
cliff=4.0.0=pyhd8ed1ab_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cmaes=0.8.2=pyh44b312d_0
cmd2=2.4.2=py310hff52083_2
colorama=0.4.6=pyhd8ed1ab_0
colorlog=6.7.0=py310hff52083_1
cryptography=38.0.1=py310h9ce1e76_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py310h7f8727e_0
dask-core=2022.7.0=py310h06a4308_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
flatbuffers=2.0.7=h27087fc_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=hc2a2eb6_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py310h5764c6d_0
fsspec=2022.3.0=py310h06a4308_0
gast=0.4.0=pyh9f0ad1d_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
glib=2.72.1=h6239696_0
glib-tools=2.72.1=h6239696_0
google-auth=2.14.1=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
greenlet=2.0.1=py310hd8f1fbe_0
grpc-cpp=1.46.4=hbad87ad_7
grpcio=1.41.1=py310h94ab34a_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.7.0=nompi_py310h416281c_102
hdf5=1.12.2=nompi_h2386368_100
icu=58.2=hf484d3e_1000
idna=3.4=py310h06a4308_0
imagecodecs=2021.8.26=py310hecf7e94_1
imageio=2.19.3=py310h06a4308_0
importlib-metadata=4.11.4=py310hff52083_0
importlib_metadata=4.11.4=hd8ed1ab_0
importlib_resources=5.10.0=pyhd8ed1ab_0
intel-openmp=2022.1.0=h9e868ea_3769
jpeg=9e=h7f8727e_0
jxrlib=1.1=h7b6447c_2
keras=2.10.0=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py310h295c915_0
krb5=1.19.3=h3790be6_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.39=hc81fddc_0
lerc=3.0=h295c915_0
libabseil=20220623.0=cxx17_h48a1fff_5
libaec=1.0.4=he6710b0_1
libblas=3.9.0=16_linux64_openblas
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=16_linux64_openblas
libclang=10.0.1=default_hb85057a_2
libcurl=7.86.0=h7bff187_1
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglib=2.72.1=h2d90d5f_0
libgomp=12.2.0=h65d4601_19
libiconv=1.16=h7f8727e_2
liblapack=3.9.0=16_linux64_openblas
libllvm10=10.0.1=he513fc3_3
libllvm13=13.0.1=hf817b99_2
libnghttp2=1.47.0=hdcd2b5c_1
libnsl=2.0.0=h7f98852_0
libogg=1.3.5=h27cfd23_1
libopenblas=0.3.21=pthreads_h78a6416_3
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libsqlite=3.39.4=h753d276_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.4.0=hecacb30_1
libuuid=2.32.1=h7f98852_1000
libvorbis=1.3.7=h7b6447c_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzlib=1.2.13=h166bdaf_4
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py310h06a4308_0
lz4-c=1.9.3=h295c915_1
mako=1.2.3=pyhd8ed1ab_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310h5764c6d_2
matplotlib=3.5.3=py310h06a4308_0
matplotlib-base=3.5.3=py310hf590b9c_0
mkl=2022.1.0=hc2b9512_224
multidict=6.0.2=py310h5764c6d_2
munkres=1.1.4=py_0
ncurses=6.3=h27087fc_1
networkx=2.8.4=py310h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.3=py310h757a811_0
numpy=1.23.4=py310h53a5b5f_1
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1s=h166bdaf_0
opt_einsum=3.3.0=pyhd8ed1ab_1
optuna=3.0.3=pyhd8ed1ab_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.4=py310h6a678d5_0
partd=1.2.0=pyhd3eb1b0_1
pbr=5.11.0=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pick=2.0.2=py310hff52083_1
pillow=9.2.0=py310hace64e9_1
pip=22.3.1=pyhd8ed1ab_0
ply=3.11=py_1
prettytable=3.4.1=pyhd8ed1ab_0
prompt-toolkit=3.0.32=pyha770c72_0
prompt_toolkit=3.0.32=hd8ed1ab_0
protobuf=3.20.1=py310hd8f1fbe_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyperclip=1.8.2=pyhd8ed1ab_2
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.6=h582c2e5_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.1=cpu_py310h75c9ab6_0
pytz=2022.1=py310h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.3.0=py310h7f8727e_0
pyyaml=6.0=py310h5764c6d_5
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
re2=2022.06.01=h27087fc_0
readline=8.1.2=h0f457ee_0
requests=2.28.1=py310h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-image=0.19.2=py310h00e6091_0
scipy=1.8.1=py310hdfbd76f_2
setuptools=65.5.1=pyhd8ed1ab_0
sip=6.6.2=py310hd8f1fbe_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
snappy=1.1.9=h295c915_0
sqlalchemy=1.4.43=py310h5764c6d_0
sqlite=3.39.3=h5082296_0
stevedore=4.1.0=pyhd8ed1ab_0
tensorboard=2.10.0=py310h06a4308_0
tensorboard-data-server=0.6.0=py310h597c629_3
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.10.0=mkl_py310h24f4fea_0
tensorflow-base=2.10.0=mkl_py310hb9daa73_0
tensorflow-estimator=2.10.0=py310h06a4308_0
termcolor=2.1.0=pyhd8ed1ab_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.2=pyhd3eb1b0_0
torchvision=0.13.0=cpu_py310h9015461_0
tornado=6.2=py310h5eee18b_0
tqdm=4.64.1=pyhd8ed1ab_0
typing_extensions=4.3.0=py310h06a4308_0
tzdata=2022f=h191b570_0
urllib3=1.26.12=py310h06a4308_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.38.3=pyhd8ed1ab_0
wrapt=1.14.1=py310h5764c6d_1
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yarl=1.8.1=py310h5764c6d_0
zfp=0.5.5=h295c915_6
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=ha4553b6_0
